refactor(ui): split PropertyInspector into multiple files
